Move init manifest generation
Previously we relied on needing to execute multiple commands to put the full manifests back together, this provides a more consistent behavior.
